requirements.txt
google-genai
requests
pillow
numpy

